requirements.txt
asgiref==3.8.1
certifi==2024.8.30
Django==4.2
django-redis==5.4.0
gunicorn==23.0.0
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
redis==5.0.8
sentry-sdk==2.15.0
sqlparse==0.5.1
urllib3==2.2.3
# micropython-imghdr
# psycopg